// web/src/api/generated/model/sittingIn.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TCFL Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { SubjectIn } from './subjectIn';

export interface SittingIn {
  key?: string | null;
  /** @pattern ^(O|A)$ */
  level: string;
  /** @pattern ^(JUNE|NOVEMBER)$ */
  session: string;
  /**
     * @minimum 1980
     * @maximum 2100
     */
  year: number;
  /** @pattern ^\d{6}$ */
  centre_number: string;
  /** @pattern ^\d{4}$ */
  candidate_number: string;
  candidate_name?: string | null;
  /**
     * @minItems 1
     * @maxItems 20
     */
  subjects: SubjectIn[];
}