// web/src/api/generated/model/subjectOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TCFL Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Crop } from './crop';

export interface SubjectOut {
  code: string | null;
  name: string;
  grade: string | null;
  check: boolean;
  read_as: string | null;
  crop: Crop | null;
}